fixed manuel's bug

+++ crop_row_detector_cpp/src/CropRowDetector.h
@@ -30,7 +30,7 @@ class CropRowDetector{
 
     std::vector<std::pair<int, int>> template_matching(cv::Mat &Intensity,
                                                        int d_min, int n_samples_per_octave, int n_octaves,
-                                                       int positive_pulse_width, int negative_pulse_width,
+                                                       double positive_pulse_width, double negative_pulse_width,
                                                        int window_width, int center_of_image_row
         );
         // std::pair<int, int> CropRowDetector::find_optimal_x(std::vector<int> f);
